update test suite with debounce and throttle tests
